Fix temperature type from Int to Double

+++ app/src/main/java/it/unibo/preh_frontend/model/VitalParametersData.kt
@@ -17,7 +17,7 @@ data class VitalParametersData(
     var pupilleDx: Int,
     var fotoreagenteSx: Boolean,
     var fotoreagenteDx: Boolean,
-    var temperature: Int,
+    var temperature: Double,
 
     var event: String = "Inseriti Parametri Vitali"
 ) : PreHData("VitalParametersData", event), Serializable
